add Bar
added Bar that displays a bar graph/chart. The values are set similarly to Line. visually, each value is a bar (rectangle), each with equal bottom edge and width but with varying top edge representing the value. the bars are separate so can be displayed separated (default). the bar separation width (distance) can be specified (defaults to 10) as can the minimum bar width (defaults to 1). note that the bar width is automatic based on the number of values, the bar width (its size's x value), and the separations; they fit the bar's overall size. the colour can be set; all bars are the same colour.

note that at least two values must be present for Bar to display anything.