add Function
added Function graph that plots a function x against y (i.e. y = f(x))

the "y=" function can be specified using the same data type as the graph.

the range of the graph can also be given (min/max x/y values).

the "number of steps" can be specified. this is the number of points along "x" that is used for plotting the graph. if this is zero, the graph's width (size's x value) is used instead. note that this is intended for use in co-ordinate systems where a single unit is useful e.g. unit per pixel.

an axis is also shown and its value position can be given. the axis can have its colour given and can also be hidden.